+++ tests/test_offline.py
@@ -8,6 +8,7 @@ USERNAME = "blota"
 password = get_password(USERNAME, "flexilims")
 PROJECT_ID = "606df1ac08df4d77c72c9aa4"  # <- test_api project
 MOUSE_ID = "6094f7212597df357fa24a8c"
+JSON_FILE = "/Volumes/lab-znamenskiyp/home/shared/projects/hey2_3d-vision_foodres_20220101/flexlims_json_version.yml"
 
 
 @pytest.mark.slow
@@ -27,3 +28,18 @@ def test_download_database(tmp_path):
 
     reloaded_data = yaml.load(open(tmp_path / "test.json"), Loader=yaml.SafeLoader)
     assert reloaded_data == json_data
+
+    def test_origin(parent):
+        for child in parent["children"].values():
+            assert child["origin_id"] == parent["id"]
+            test_origin(child)
+
+    test_origin(json_data["test_mouse"])
+
+
+def test_format_df():
+    import pandas as pd
+    from flexilims.offline import OffFlexilims
+
+    sess = OffFlexilims(JSON_FILE)
+    assert isinstance(sess._dataframe, pd.DataFrame)